Version 1.0.44
Fixed OnClipboardChange to work even when the script is displaying a MsgBox or other dialog.

Fixed FileCreateDir not to report ErrorLevel 1 when the specified directory ends in "\". [thanks Wesley]

Fixed hotkeys Control::, Shift::, and Alt:: to fire on release of the key rather than press-down (broken by v1.0.41).

Fixed loading of icon #1 from nonstandard file types such as .bin.

Fixed ListView creation to obey the text color set by "Gui Font".

Changed and fixed the hotkey tilde prefix (~) so that hotkeys like ~Esc and Esc are considered duplicates; that is, a script cannot define both ~Esc and Esc unless they are under different #IfWin criteria (in which case ~Esc and Esc have been fixed to work properly). [thanks jballi]

Changed and improved ListViews: 1) a new option WantF2 is in effect by default, which causes an F2 keystroke to edit the current row (when not ReadOnly); 2) added LV_Modify(Row, "Vis") to scroll an item into view; 3) drag notifications occur even without AltSubmit; 4) item-change notifications have A_EventInfo set to the row number; 5) ErrorLevel no longer mirrors A_EventInfo in cases where it was never documented to do so (same for GuiContextMenu).

Improved Edit controls to select all text in response to Control-A. This can be disabled via -WantCtrlA.

Added new GUI control types TreeView and StatusBar.

+++ Source/script_expression.cpp
@@ -1489,8 +1489,8 @@ skip_abort_udf:
 				// an integer above.
 				if (right_int64 < 0 || right_int64 > UINT_MAX)
 					this_token.value_int64 = ~right_int64;
-				else // See comments at TRANS_CMD_BITNOT for why it's done this way.
-					this_token.value_int64 = ~((DWORD)right_int64);
+				else // See comments at TRANS_CMD_BITNOT for why it's done this way:
+					this_token.value_int64 = (size_t)~(DWORD)right_int64; // Casting this way avoids compiler warning.
 			}
 			this_token.symbol = SYM_INTEGER; // Must be done only after its old value was used above. v1.0.36.07: Fixed to be SYM_INTEGER vs. right_is_number for SYM_BITNOT.
 			break;
